\name{camargo2001}
\alias{camargo2001}
\docType{data}
\title{ Data Frame Containing Sample Weight (Grams) of each Mesh Size (phi units) }
\description{  The first line is the mesh size (phi units). The row names are the sample names  }
\usage{data(camargo2001)}
\format{
  A data frame with 15 observations on the following 17 variables.
}
\references{
- Camargo, M.G. (2001)  \emph{Analise experimental das interacoes biologicas da macrofauna e meiofauna bentica de manguezais das baias de paranagua e antonina - PR.}, PhD Thesis. UFPR. Brazil.\cr
}

\examples{
data(camargo2001)
}
\keyword{datasets}
