\name{sed.metric}
\alias{sed.metric}
\docType{data}
\title{Data Frame Containing Sample Weight (Grams) of each Mesh Size (micrometers)}
\description{ The first line is the mesh size (micrometers). The row names are the sample names.\cr
Each ten samples represent a different environment (in order: Rivers, Tidal Flats, Subtidal and Beaches)
}
\usage{data(sed.metric)}
\format{
  A data frame with 41 observations on the following 17 variables.
}
\examples{
data(sed.metric)
}
\keyword{datasets}
