\name{sed.phi}
\alias{sed.phi}
\docType{data}
\title{ Data Frame Containing Sample Weight (Grams) of each Mesh Size (phi units) }
\description{ The first line is the mesh size (phi units). The row names are the sample names.\cr
Each ten samples represent a different environment (in order: Rivers, Tidal Flats, Subtidal and Beaches)
}
\usage{data(sed.phi)}
\format{
  A data frame with 41 observations on the following 17 variables.
}
\examples{
data(sed.phi)
}
\keyword{datasets}
