\name{camargo2001}
\alias{camargo2001}
\docType{data}
\title{Dataframe Containing Sample Weight (Grams) of each Mesh Size (phi units)}
\description{  The first line is the mesh size (phi units). The row names are the sample names  }
\usage{data(camargo2001)}
\format{
  A dataframe with 36 observations on the following 17 variables.
  \describe{
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{V7}}{a numeric vector}
    \item{\code{V8}}{a numeric vector}
    \item{\code{V9}}{a numeric vector}
    \item{\code{V10}}{a numeric vector}
    \item{\code{V11}}{a numeric vector}
    \item{\code{V12}}{a numeric vector}
    \item{\code{V13}}{a numeric vector}
    \item{\code{V14}}{a numeric vector}
    \item{\code{V15}}{a numeric vector}
    \item{\code{V16}}{a numeric vector}
    \item{\code{V17}}{a numeric vector}
    \item{\code{V18}}{a numeric vector}
  }
}
\references{
- Camargo, M.G. (2001)  \emph{Analise experimental das interacoes biologicas da macrofauna e meiofauna bentica de manguezais das baias de paranagua e antonina - PR.}, PhD Thesis. UFPR. Brazil.\cr
}

\examples{
data(camargo2001)
}
\keyword{datasets}
